[doxyrest_sphinx] modify cref role to allow using it against groups
group names usually contain spaces, which makes it impossible to use cref as is
(need to replace spaces with '-'). also, when space is found, it clearly means
that target is NOT a code entity; therefore, don't use nodes.literal for it.

end result -- now it's possible to refer to groups like this (given cref is the
default role):

Please check `Base Definitions` for more details.