[c_frame] removed argstring formatting from getTypedefDeclString
mostly, this concerns function pointer typedefs. however, the former implementation simply produced incorrect results in the most general case, so i decided to revert everything to the simple concatenation of argString.

jancy compiler is populating paramArray, so the result should stay exactly the same as before; for C/C++ the output will change to the one produced for variables.

proper analysis and formatting should be added so function pointer typedefs look polished, but for now it's acceptable -- correctness of the output is way more important than the polished look.