Add graph output to lua frame base
Supports 3 out of 4 graph types supported by doxygen. For those 3
graph types there are sphinx extensions available to render them:

* dot -> sphinx.ext.graphviz
* plantuml -> sphinxcontrib.plantuml
* msc -> sphinxcontrib.mscgen

If the required extension isn't available, the graph code will not be
given in the html output which is also a plus IMO, as it normally
clutters the descritpion.

It seems there is no extension available for dia. Also missing currently
is support for graph file references, as well as passthrough of
captions.